carssite/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
Django==4.1.3
django-debug-toolbar==3.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.3.0
pycodestyle==2.9.1
sqlparse==0.4.3
tomli==2.0.1
